Prevent list number resetting
Within the Markdown renderer in GitHub's edit interface, including this text on a separate line does not affect the list numbering, so that the following item is properly numbered as 5. But within the Markdown renderer on the site, this is not so—the list goes 1, 2, 3, 4, 1. Combining these two lines into a single line corrects this.